refacotr: userInfo 파일 분리 및 (getBrowser, getOs, getIsMobile, getResoulition, getUserCountry)생성, heartbeat (load 이벤트 삭제,), utils(getDevice userInfo로 변경)
